feat(stream): event metadata
- add possibility to provice metadata for an event
- the metadata field is a variable length field in the header of an
  event.
- commit also documents header descriptor, rearranges header fields and
  "downgrades" some fields of type uint64 to uint32 and reserves a
  version field in header

related to #38

+++ logstreams/src/main/java/org/camunda/tngp/logstreams/LogStreams.java
@@ -17,7 +17,7 @@ public class LogStreams
         return new FsSnapshotStorageBuilder(rootPath);
     }
 
-    public static StreamProcessorBuilder createStreamProcessor(String name, long id, StreamProcessor streamProcessor)
+    public static StreamProcessorBuilder createStreamProcessor(String name, int id, StreamProcessor streamProcessor)
     {
         return new StreamProcessorBuilder(id, name, streamProcessor);
     }
